Zeus: Master of Olympus — Syntax update
* Update script to 2.11 syntax
* Set WINE virtual desktop resolution to current desktop resolution
* Use generic dependencies

+++ play.it-2/games/play-zeus-master-of-olympus.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -30,18 +30,16 @@ set -o errexit
 
 ###
 # Zeus: Master of Olympus
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20190707.1
 
 # Set game-specific variables
 
 GAME_ID='zeus-master-of-olympus'
 GAME_NAME='Zeus: Master of Olympus'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='setup_zeus_and_poseidon_2.1.0.10.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/zeus_poseidon'
@@ -49,27 +47,26 @@ ARCHIVE_GOG_MD5='f26f9ed5ecaa4e58fca64acb88255107'
 ARCHIVE_GOG_SIZE='800000'
 ARCHIVE_GOG_VERSION='2.1-gog2.1.0.10'
 
-ARCHIVE_DOC1_DATA_PATH='tmp'
-ARCHIVE_DOC1_DATA_FILES='./*.txt'
+ARCHIVE_DOC0_DATA_PATH='tmp'
+ARCHIVE_DOC0_DATA_FILES='*.txt'
 
-ARCHIVE_DOC2_DATA_PATH='app'
-ARCHIVE_DOC2_DATA_FILES='./*.txt ./*.pdf'
+ARCHIVE_DOC1_DATA_PATH='app'
+ARCHIVE_DOC1_DATA_FILES='*.txt *.pdf'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./*.asi ./*.exe ./*.ini ./*.m3d ./binkw32.dll ./ijl10.dll ./mss32.dll'
+ARCHIVE_GAME_BIN_FILES='*.asi *.exe *.ini *.m3d binkw32.dll ijl10.dll mss32.dll'
 
 ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./*.eng ./*.inf ./poseidon.ico ./zeus.ico ./adventures ./audio ./binks ./data ./model'
+ARCHIVE_GAME_DATA_FILES='*.eng *.inf poseidon.ico zeus.ico adventures audio binks data model'
 
 CONFIG_FILES='./*.ini'
 DATA_DIRS='./save'
 
-APP_WINETRICKS='vd=1024x768'
+APP_WINETRICKS="vd=\$(xrandr|awk '/\\*/ {print \$1}')"
 
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='zeus.exe'
 APP_MAIN_ICON='poseidon.ico'
-APP_MAIN_ICON_RES='16 32'
 
 PACKAGES_LIST='PKG_DATA PKG_BIN'
 
@@ -77,47 +74,51 @@ PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_DATA_ID, wine32-development | wine32 | wine-bin | wine-i386 | wine-staging-i386, wine:amd64 | wine, winetricks"
-PKG_BIN_DEPS_ARCH="$PKG_DATA_ID wine winetricks"
+PKG_BIN_DEPS="$PKG_DATA_ID wine winetricks"
 
 # Load common functions
 
-target_version='2.1'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC1_${PKG#PKG_}" "$PATH_DOC"
-	organize_data "DOC2_${PKG#PKG_}" "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
+# Extract icons
 
 PKG='PKG_DATA'
-extract_and_sort_icons_from 'APP_MAIN'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
